refactor(musefs-core): split facade refresh engine into facade/refresh.rs
The single `impl Musefs` block mixed handle/fd lifecycle, the serve/read
path, and the ~600-line incremental-refresh/rebuild engine. Move the engine
(`render_one`..`expire_refresh_backoff_for_test`, plus `RefreshGuard`,
`retry_backoff_for`, `IncrementalOutcome`) into `facade/refresh.rs` as a
second `impl Musefs`, and extract the inline test module to `facade/tests.rs`
via a plain `#[cfg(test)] mod tests;`. facade.rs drops 2487 -> ~695 lines.

The six helpers reached across the new module boundary (`build_full`,
`render_entries`, `order_entries`, `force_full_rebuild`,
`force_version_mismatches_for_test`, `retry_backoff_for`) are bumped to
`pub(crate)`; `refresh.rs` uses explicit imports because the workspace's
pedantic `wildcard_imports` lint forbids `use super::*` outside test modules.

Addresses #445 (concern split) and #444 (test extraction) for facade.
